+++ NEWS.md
@@ -1,5 +1,18 @@
 # forestr 0.0.0.9000
 
+TODO
+
+* refactor `to_id_rowcol()`
+
+
+
+
+DONE
+
+* Renamed
+* Added assertions
+* Improved documentation
+
 * At the workshop in Puerto Rico (2017) the priority was set (private table of functions' priority https://goo.gl/ijZcZL).